refactor(runtime-update): move the exclusive-access wrapper into the plugin
`ExclusiveRuntimePlugin` is a concrete locking policy built on
`tokio::sync::RwLock`, not a contract, so it did not belong in a crate whose
purpose is to hold contracts. Its presence there also obliged `cda-interfaces`
to carry a tokio synchronisation dependency for something no interface needs.

`with_exclusive_access` was a provided method on `RuntimeFilesUpdatePlugin`,
which forced every implementation to inherit one particular concurrency
strategy as part of implementing the trait. It becomes a blanket-implemented
extension trait instead: still one call to opt in, but now visibly a choice. An
implementation that already serialises internally, or that wants different
semantics, simply does not use it.

While moving, the doc records what the wrapper actually arbitrates. It is not
about competing plugins - exactly one update plugin is ever registered - but
about concurrent HTTP requests against that one plugin, which would otherwise
interleave on the same staging directory.
